% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mozart.R
\docType{data}
\name{mozart}
\alias{mozart}
\title{Data of studies on the Mozart effect for meta-analysis}
\format{A data frame with 38 rows and 6 variables:
\describe{
  \item{study_name}{short name of each study}
  \item{n}{sample size of each study}
  \item{d}{observed
  effect size (Cohen \emph{d}). Positive values correspond to higher mean
  performance in the group hearing the Mozart sonata}
  \item{se}{standard
  error of observed effect size}
  \item{unpublished}{dichotomous moderator variable: Was the
  study unpublished?}
  \item{rr_lab}{dichotomous moderator variable: Was the study
  conducted in the lab of authors Rauscher or Rideout?}
  }}
\usage{
mozart
}
\description{
A dataset consisting of 38 empirical studies used in the meta-analysis of
Pietschnig, Voracek, and Formann (2010) on the Mozart effect. Each study compared the spatial
task performance of participants after hearing the first movement "allegro
con spirito" of the Mozart sonata for two pianos in D major (KV 448) with
the spatial task performance of participants who were exposed to a non-musical
stimulus or no simulus at all.
}
\references{
Pietschnig, J., Voracek, M., & Formann, A. K. (2010). Mozart
  effect-Shmozart effect: A meta-analysis. \emph{Intelligence}, \emph{38}, 314-323.
}
\keyword{datasets}
