% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem_functions.R
\name{rem_effect}
\alias{rem_effect}
\title{Internal helper function of rainforest(): Summary effect (REM)}
\usage{
rem_effect(es, se)
}
\description{
Computes the meta-analytic summary effect(s) for the random effects model (for method = "REM")
}
\keyword{internal}

