% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_powerfun.R
\name{viz_powerfun}
\alias{viz_powerfun}
\title{Power enhanced funnel plot}
\usage{
viz_powerfun(x, y_axis = "se", method = "FE", true_effect = NULL,
  power_contours = "continuous", contours = TRUE, sig_contours = FALSE,
  detail_level = 1, text_size = 3, point_size = 2, xlab = "Effect",
  ylab = NULL, x_trans_function = NULL, x_breaks = NULL)
}
\arguments{
\item{x}{data.frame or matrix with the effect sizes of all studies (e.g.,
correlations, log odds ratios, or Cohen \emph{d}) in the first column and their
respective standard errors in the second column. Alternatively, x can be the
output object of function \code{\link[metafor]{rma.uni}} from package
\pkg{metafor}; then effect sizes and standard errors are extracted from \code{x}.}

\item{y_axis}{character string indicating which y axis should be used in the funnel plot. Available options are "se" (default) for
standard error and "precision" for the reciprocal of the standard error.}

\item{method}{character string indicating the method used to compute the meta-analytic summary effect and, for a random effects model,
the between-study variance \eqn{\tau^2}{tau squared}. Can be any method argument from \code{\link[metafor]{rma.uni}}
(e.g., "FE" for the fixed effect model, or "DL" for the random effects model using the DerSimonian-Laird method to estimate \eqn{\tau^2}{tau squared}).
Used for \code{contours}, and \code{addev_contours}.}

\item{true_effect}{numeric scalar. Which true effect should be assumed for power caluclations? By default the meta-analytic summary effect is used.}

\item{power_contours}{character string specifying how different power regions are plotted. Can be eiter "continuous"
(default) or "discrete".}

\item{contours}{logical scalar indicating if classic funnel plot confidence contours and the summary effect
should be displayed.}

\item{sig_contours}{logical scalar. Should significance contours be drawn (at the 0.05 or 0.01 level using a Wald test)?}

\item{detail_level}{numeric scalar. Allows to increase or decrease the plotting detail of contours. Values can be chosen between 0.1 and 10. Default is 1.}

\item{text_size}{numeric value. Size of text in the funnel plot. Default is 3.}

\item{point_size}{numeric value. Size of the study points in the funnel plot. Default is 2.}

\item{xlab}{character string specifying the label of the x axis.}

\item{ylab}{character string specifying the label of the y axis.}

\item{x_trans_function}{function to transform the labels of the x axis. Common uses are to transform
log-odds-ratios or log-risk-ratios with \code{exp} to their original scale (odds ratios and risk ratios), or Fisher's z values
back to correlation coefficients using \code{tanh}.}

\item{x_breaks}{numeric vector of values for the breaks on the x-axis. When used in tandem with \code{x_trans_function}
the supplied values should be not yet transformed.}
}
\value{
A funnel plot with power contours is created using ggplot2.
}
\description{
Creates a funnel plot with power regions.
}
\details{
The funnel plot is a widely used diagnostic plot in meta-analysis to assess small study effects
and in particular publication bias. Small study effects can be defined as association of effect size and standard error,
such that small studies (with large standard errors) tend to report larger effect sizes on average than larger studies
with preciser estimates. This is widely seen as indicator for potential publication bias, because small studies need to observe
larger effect sizes to obtain significant results. The power of a study is a function of its standard error (for a fixed true effect and significance level).
Therefore it can be useful to assess if indeed low powered studies tend to report larger and significant results. For this purpose,
the power corresponding to a given standard error is computed by using a two-sided Wald test and (by default) the meta-analytic summary effect as true effect.
Colored regions of different power levels and a second axis with study level power is shown within the funnel plot.
}
\examples{
library(metaviz)
# Create a funnel plot using confidence and significance contours
viz_powerfun(x = mozart[1:10, c("d", "se")])
}
\author{
Michael Kossmeier* <michael.kossmeier@univie.ac.at>

Ulrich S. Tran* <ulrich.tran@univie.ac.at>

Martin Voracek* <martin.voracek@univie.ac.at>

*Department of Basic Psychological Research and Research Methods, School of Psychology, University of Vienna
}
