% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_forest_internal.R
\name{viz_thickforest_internal}
\alias{viz_thickforest_internal}
\title{Internal helper function of viz_thickforest and viz_forest to create a thick forest plot}
\usage{
viz_thickforest_internal(plotdata, madata, type = "standard",
  study_labels = NULL, summary_label = NULL, study_table = NULL,
  summary_table = NULL, annotate_CI = FALSE, confidence_level = 0.95,
  col = "Blues", summary_col = "black", tick_col = "firebrick",
  text_size = 3, xlab = "Effect", x_limit = NULL,
  x_trans_function = NULL, x_breaks = NULL)
}
\description{
Creates a thick forest plot. Called by viz_thickforest and viz_forest for type = "thick"
}
\keyword{internal}
