% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tlma_studyinfo.R
\name{viz_tlma_studyinfo}
\alias{viz_tlma_studyinfo}
\title{A dataframe containing information on the study effects of a three-level meta-analysis}
\usage{
viz_tlma_studyinfo(x, confidence_level_ci = 0.95, ordered = FALSE)
}
\arguments{
\item{x}{metafor rma.mv object}

\item{confidence_level_ci}{numeric confidence level for the confidence intervals of the study effects}

\item{ordered}{orders the data by effect size when TRUE}
}
\value{
A dataframe containing statistical information about the study effects of a three-level meta-analysis is created.
}
\description{
Provides a dataframe with information on the study effects of a three-level meta-analysis.
}
\details{
The function viz_tlma_studyinfo creates a dataframe with information pertaining to the study effects of
a three-level meta-analysis. The study effects are the result of random-effects models conducted on the effects
contained within each respective study. The dataframe contains the number of effects originating from the study,
the study effect with its standard error and confidence intervals, and the weight of the study within the three-level
meta-analysis.
}
\examples{

if (requireNamespace("psymetadata", quietly = TRUE)) {
  # Get wibbelink2017 data
  testdata <- psymetadata::wibbelink2017


  # Calculate the three-level meta-analytic model
  testmodel <- metafor::rma.mv(yi,
                              vi,
                              random = ~ 1 | study_id/es_id,
                              tdist = TRUE,
                              data = testdata,
                              method = "REML")

  # Create a table with study-level information
  viz_tlma_studyinfo(testmodel)
}
}
\author{
Verena Pilar <verena.pilar@outlook.com>
}
