% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wineq_forest_internal.R
\name{internal_wineq_forest_thick}
\alias{internal_wineq_forest_thick}
\title{Internal helper function of wineq_forest to create a thick wineq forest plot.}
\usage{
internal_wineq_forest_thick(
  plotdata,
  madata,
  summary_line = NULL,
  method = "REML",
  study_labels = NULL,
  summary_label_FE = NULL,
  summary_label_REML = NULL,
  study_table = NULL,
  summary_table = NULL,
  annotate_CI = FALSE,
  confidence_level = 0.95,
  col = NULL,
  summary_col_FE = "steelblue1",
  summary_col_REML = "steelblue4",
  tick_col = "firebrick",
  errorbar_col = TRUE,
  text_size = 3,
  xlab = "Effect",
  x_limit = NULL,
  x_trans_function = NULL,
  x_breaks = NULL
)
}
\description{
Creates a thick wineq forest plot. Called by wineq_forest for variant = "thick"
Note: This function was developed on the basis of the viz_forest_internal code which was created by Michael Kossmeier.
}
\keyword{internal}
