% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_forest.R
\name{viz_forest}
\alias{viz_forest}
\title{Forest plot variants for meta-analyses}
\usage{
viz_forest(x, group = NULL, type = c("classic", "rain", "thick"),
  summary_symbol = TRUE, method = "FE", study_labels = NULL,
  summary_label = NULL, study_table = NULL, summary_table = NULL,
  table_headers = NULL, annotate_CI = FALSE, confidence_level = 0.95,
  col = "Blues", text_size = 3, xlab = "Effect", x_limit = NULL,
  x_trans_function = NULL, x_breaks = NULL, ...)
}
\arguments{
\item{x}{data.frame or matrix with the effect sizes of all studies (e.g.,
correlations, log odds ratios, or Cohen \emph{d}) in the first column and their
respective standard errors in the second column. Alternatively, x can be the
output object of function \code{\link[metafor]{rma.uni}} from package
\pkg{metafor}.}

\item{group}{factor indicating the subgroup of each study to plot a subgroup forest plot. Has to be in the same order than \code{x}.}

\item{type}{character indicating the forest plot variant that should be plotted. Can be "rain" for rainforest plot,
"thick" for a thick forest plot, or "classic" for a traditional forest plot. Default is "classic".}

\item{summary_symbol}{logical scalar. Should the meta-analytic summary result(s) be computed (using \code{method}) and shown?}

\item{method}{Which method should be used to compute the study weights and summary effect(s)?
Can be any method argument from \code{\link[metafor]{rma.uni}}
(e.g., "FE" for the fixed effect model, or "DL" for the random effects model using the
DerSimonian-Laird method to estimate \eqn{\tau^2}{tau squared}).}

\item{study_labels}{a vector with names/identifiers to annotate each study in the forest plot.
Has to be in the same order than \code{x}.}

\item{summary_label}{a name to annotate the summary effect. If a subgroup
analysis is plotted, \code{summary_label} should be a vector with a name for each
subgroup summary effect, arranged in the order of the levels of \code{group}.}

\item{study_table}{a data.frame with addtional study-level variables which should be shown in an aligned table.
Has to be in the same order than \code{x}. See vignette('metaviz').}

\item{summary_table}{a data.frame with addtional summary-level information shown in an aligned table.
If \code{group} is supplied, \code{summary_table} must have a row for each subgroup
summary effect, arranged in the order of the levels of \code{group}. See vignette('metaviz').}

\item{table_headers}{character vector. Headers for each column of the aligned table if \code{study_table} and/or \code{summary_table} is supplied.
By default the names of \code{study_table}.}

\item{annotate_CI}{logical scalar. Should the effect size and confidence interval values be annotated?}

\item{confidence_level}{the confidence level for the plotted confidence bars.}

\item{col}{character specifying the main color for plotting. For type = "rain" must be one of the follwoing palettes from package
\pkg{RColorBrewer}: "Blues", "Greys", "Oranges", "Greens", "Reds", or "Purples".}

\item{text_size}{numeric value. Values larger than 1 lead to larger text size,
values smaller than 1 to smaller text size than the default.}

\item{xlab}{character label of the x axis. Also used for the header of the aligned table if \code{annotate_CI} is TRUE.}

\item{x_limit}{numeric vector of length 2 with the limits (minimum, maximum) of the x axis.}

\item{x_trans_function}{function to transform the labels of the x axis. Common uses are to transform
log-odds-ratios or log-risk-ratios with \code{exp} to their original scale (odds ratios and risk ratios), or Fisher's z values
back to correlation coefficents using \code{tanh}. See vignette('metaviz').}

\item{x_breaks}{numeric vector of values for the breaks on the x-axis. When used in tandem with \code{x_trans_function}
the supplied values should be not yet transformed.}

\item{...}{furhter arguments passed to \code{\link[metaviz]{viz_rainforest}} for type = "rain", or
\code{\link[metaviz]{viz_thickforest}} for type = "thick".}
}
\value{
A forest plot is created using ggplot2.
}
\description{
Creates a rainforest, thick forest, or classic forest plot.
}
\details{
The forest plot is probably the most widley used display to visualize meta-analytc results.
The function \code{viz_forest} creates visually appealing forest plots using ggplot2. Many options
to flexibly customize the visual appearence and statistical information displayed are provided. In addtion,
rainforest plots as well as the thick forest plots can be created, two variants and enhancements of the
classical forest plot recently proposed by Schild and Voracek (2015). For furhter details see the documentation of
\code{\link[metaviz]{viz_rainforest}}, and \code{\link[metaviz]{viz_thickforest}}.
}
\examples{
library(metaviz)
# Plotting a rainforest plot using the mozart data (for details, see help(mozart)):
viz_forest(x = mozart[, c("d", "se")],
study_labels = mozart[, "study_name"], xlab = "Cohen d")

# Subgroup analysis of published and unpublished studies with a classic forest plot
viz_forest(x = mozart[, c("d", "se")], study_labels = mozart[, "study_name"],
type = "thick", summary_label = c("Summary (published)", "Summary (unpublished)"),
group = mozart[, "unpublished"], xlab = "Cohen d")

# Thick forest plot with additional information in aligned tables. Log risk
# ratios are labeled in their original metric (risk ratios) on the x axis.
viz_forest(x = exrehab[, c("logrr", "logrr_se")], type = "rain",
study_labels = exrehab[, "study_name"],
annotate_CI = TRUE, xlab = "RR", x_trans_function = exp,
study_table = data.frame(
eventsT = paste(exrehab$ai, "/", exrehab$ai + exrehab$bi, sep = ""),
eventsC = paste(exrehab$ci, "/", exrehab$ci + exrehab$di, sep = "")),
summary_table = data.frame(
eventsT = paste(sum(exrehab$ai), "/", sum(exrehab$ai + exrehab$bi), sep = ""),
eventsC = paste(sum(exrehab$ci), "/", sum(exrehab$ci + exrehab$di), sep = "")))
}
\references{
Schild, A. H., & Voracek, M. (2015). Finding your way out of the
 forest without a trail of bread crumbs: Development and evaluation of two
 novel displays of forest plots. \emph{Research Synthesis Methods}, \emph{6},
 74-86.
}
\author{
Michael Kossmeier* <michael.kossmeier@univie.ac.at>

Ulrich S. Tran* <ulrich.tran@univie.ac.at>

Martin Voracek* <martin.voracek@univie.ac.at>

*Department of Basic Psychological Research and Research Methods, School of Psychology, University of Vienna
}
