% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tlma_studyinfo.R
\name{viz_tlma_studyinfo}
\alias{viz_tlma_studyinfo}
\title{A dataframe containing information on the study effects of a three-level meta-analysis}
\usage{
viz_tlma_studyinfo(x, study_ID, effect_ID, confidence_level_ci = 0.95)
}
\arguments{
\item{x}{metafor rma.mv object}

\item{study_ID}{grouping variable for the study-level of the three-level meta-analysis that was used in the rma.mv model}

\item{effect_ID}{grouping variable for the effect-level of the three-level meta-analysis that was used in the rma.mv model}

\item{confidence_level_ci}{numeric confidence level for the confidence intervals of the study effects}
}
\value{
A dataframe containing statistical information about the study effects of a three-level meta-analysis is created.
}
\description{
Provides a dataframe with information on the study effects of a three-level meta-analysis.
}
\details{
The function viz_tlma_studyinfo creates a dataframe with information pertaining to the study effects of
a three-level meta-analysis. The study effects are the result of random-effects models conducted on the effects
contained within each respective study. The dataframe contains the number of effects originating from the study,
the study effect with its standard error and confidence intervals, and the weight of the study within the three-level
meta-analysis.
}
\examples{
library(metafor)
library(psymetadata)
# Extract wibbelink2017 data
testdata <- wibbelink2017

# Determine level 1 and 2 grouping variables for the three-level meta-analysis
# study IDs = level 2
ID <- testdata$study_id
# effect IDs = level 1
ID2 <- testdata$es_id

# Calculate the three-level meta-analytic model
testmodel <- rma.mv(yi,
                    vi,
                    random = ~ 1 | ID/ID2,
                    tdist = TRUE,
                    data = testdata,
                    method = "REML")

# Create a table with study-level information
viz_tlma_studyinfo(testmodel, ID, ID2)
}
\author{
Verena Pilar <verena.pilar@univie.ac.at>
}
