% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wineq_gini.R
\name{wineq_gini}
\alias{wineq_gini}
\title{Lorenz curves for the comparison of study weight concentration of fixed-effect and random-effects models}
\usage{
wineq_gini(x, type = "FEM_REM", col = FALSE, tables = TRUE, seed = NULL)
}
\arguments{
\item{x}{metafor rma.uni object conducted with method “FE” or “REML” (the chosen method of the input model makes no difference in the resulting plot)}

\item{type}{determines which Lorenz curves are shown. The default “FEM_REM” shows Lorenz curves for both the fixed-effect and random-effects models,
“FEM” shows only the fixed-effect Lorenz curve and “REM” shows only the random-effects Lorenz curve.}

\item{col}{boolean argument that determines whether Lorenz curves are colored by model or not.}

\item{tables}{boolean argument that determines whether a large table with lots of statistical information is plotted below the coordinate system (“TRUE”)
or whether a small table with information only on the Gini indices is shown within the coordinate system (“FALSE”)}

\item{seed}{numeric argument that is used to set a random seed in order to provide reproducible bootstrapped Gini index
confidence intervals. If seed == NULL the Gini index confidence intervals will vary slightly between instances of plotting.}
}
\value{
Two Lorenz curves are plotted in the same coordinate system accompanied by a table of statistical information.
}
\description{
Creates two Lorenz curves within the same coordinate system that correspond to the study weight concentration
in a fixed-effect and a random-effects model, respectively, based on the same data.
}
\details{
The function wineq_gini creates a plot containing two Lorenz curves (Lorenz, 1905) which represent the concentration of weights among the studies
within a fixed-effect model and random-effects model, respectively. An adjoined table provides descriptive statistics about the study weights, as well as
Gini indices which correspond to the Lorenz curves (Gini, 1912; Tran et al, 2021). The Gini quotient quantifies the discrepancy between the two Lorenz curves,
is directly related to heterogeneity and serves as an effect size for cross-metaanalytic comparisons.
}
\examples{
library(metafor)

# Calculating a random-effects model based on the mozart data
mozart_r <- rma(yi = d,
                sei = se,
                data = mozart,
                method = "REML")

# using a rma.uni model as input
wineq_gini(x = mozart_r, seed = 123)
}
\references{
Gini, C. (1912). Variabilità e mutabilità (Variability and Mutability). C. Cuppini, Bologna, 156.

Lorenz,M.O. (1905). Methods of measuring the concentration of wealth. \emph{Pub. Am. Stat. Assoc.} 9, 209–219. https://doi.org/10.2307/2276207

Tran, U. S., Lallai, T., Gyimesi, M., Baliko, J., Ramazanova, D., & Voracek, M. (2021). Harnessing the fifth element of distributional statistics for psychological science:
A practical primer and shiny app for measures of statistical inequality and concentration.\emph{ Frontiers in Psychology}, 12, 716164.
}
\author{
Verena Pilar <verena.pilar@outlook.com>
}
